update method impl

+++ projects/library/src/main/java/com/library/users/service/UserService.java
@@ -7,8 +7,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 
+import com.library.books.model.Book;
+import com.library.books.model.BookNotFoundException;
+import com.library.users.model.CannotDeleteAdminException;
+import com.library.users.model.Role;
 import com.library.users.model.User;
 import com.library.users.model.UserAdapter;
+import com.library.users.model.UserExistsException;
 import com.library.users.model.UserNotFoundException;
 import com.library.users.repo.UserRepository;
 
@@ -23,9 +28,28 @@ public class UserService{
 	}
 	
 	public User createUser(User new_user){
+		if(checkEmail(new_user.getEmail()) || checkUserName(new_user.getUsername()))
+			throw new UserExistsException();
+		new_user.setRole(Role.USER);
 		return userRepository.save(new_user);
 	}
 	
+	private boolean checkUserName(String username) {
+		for(User user : userRepository.findAll()) {
+			if(user.getUsername().equals(username))
+				return true;
+		}
+		return false;
+	}
+
+	private boolean checkEmail(String email) {
+		for(User user : userRepository.findAll()) {
+			if(user.getEmail().equals(email))
+				return true;
+		}
+		return false;
+	}
+
 	public List<User> getUsers(){
 		return userRepository.findAll();
 	}
@@ -42,6 +66,20 @@ public class UserService{
 	public void deleteUser(Long id) {
 		User user = getUsersById(id);
 		
+		if(user.getRole().equals(Role.ADMIN))
+			throw new CannotDeleteAdminException();
+		
 		userRepository.delete(user);
+	}
+	
+	public List<Role> getRoles() {
+		return List.of(Role.values());
+	}
+	
+	public void updateUser(User user) {
+		if(userRepository.existsById(user.getId()))
+			userRepository.save(user);
+		
+		throw new UserNotFoundException(user.getId());
 	}
 }
